Move mocking of bspline to seperate file and exclude it from oclint checks.
