+++ scripts/docker/Dockerfile.lint.oclint
@@ -21,4 +21,4 @@ WORKDIR /code/build-clang-tidy
 RUN ln -s ../.oclint .
 
 # execute clang-tidy
-CMD oclint-json-compilation-database -- -enable-clang-static-analyzer -enable-global-analysis -max-priority-1=0 -max-priority-2=0 -max-priority-3=0
+CMD oclint-json-compilation-database -e .*/2d_b_spline_mocking.h -- -enable-clang-static-analyzer -enable-global-analysis -max-priority-1=0 -max-priority-2=0 -max-priority-3=0
